Collapse tile func is finished and ready
collapse is ready and adds surrounding tiles to the stack for the propagation function which just needs to be finished, the for loops for checking the tile if there is more than one tile in the cell, just needs to be finished checking over all the indivitual tiles and chekcing if the tiles are compatible with the surrounding 4 tiles if they match, just use a for loop to check every negx value of every tile with the posx and the negy with the negx, etc. might be a bit sloppy in terms of performance but levels are gonna be kept under 1000x1000 anyways so it shouldent be too bad. Also tile weights are now working and should be adjusted accoudingly in the tilevalues.json file