removed bloat code from func propagate()
Organized the current code and changed how the finished map i checked, probably a lot worse on performance but its not too slow. Tiles seem to be miss matched, and not properly placed with sockets, or sides labels properly, might still be a problem with propagate but im guessing its with the collapse function grabbing something at random before the tiles are organized, however if its ran after propagate it should not need to check the sides in the first place, and therefore be an issue with propagation either no checking properly or not updating the remaining tiles before the collapse, the delete function seems to be getting loaded up with a lot of tiles and deleting most of the tiles with 1 check, my guess is thats what causing the wrong tile placement issue. this is still a sloppy design tho, once everything is figured out and organized the v2 or v3, i forget, will be organized with more functions, mainly one that checks for each angle instead of having 4 if blocks to check every direction.