Fixed bugs with path finding, work as intended.
Path finding with the small tile set and bigger enemy is causing issues with getting around corners with the straight forward, Corridorfunnel post pressing setting, Edge centered makes weird paths however, might be settings to narrow the leaniancy with the "Center" of the pathing. There may also be a different algorithm than A-Star made.