Added weapon pick up node
can now pick up weapons and checks if you already have a weapon, but just deletes weapon node since there is no ammo system, not sure if we will add one. Still need to figure out if there will be a weapon carry limit and if so then add a way for weapons to be dropped. There is a bug if the player has no weapons and tries to pick up a weapon then the player is unable to switch to the weapon, probably due to something not setting up the array, could be an abstraction problem, could be solved by a simple check if the array is empty then to fill it with the new weapon and make it selected, also could be due to the selection counter being set wrong compensating for a weapon that is not there. Not a problem atm but should get looked into.