Added Character Sheet json file for stat changes
made a new function that runs on ready with a new var, new_run if new_run is true that indicates its a new run, obviously, otherwise will be set false for if your on later stages. A new run will make a new fresh character sheet from the default one, and hold it in a dictionary called character sheet, will need to save to a new temporarily json or resource when loading between levels or saving. This way we can keep track of modifiers and change values on the fly, right now i only have a few different stats like movement and fire spread, movement of which is the only one calculating for in the character sheet so ill have to update all the code in the player controller to account for a multiplier, as well as finishing the item manager node that keeps track of all the items. Its not attached to the character so it can be easy er to reference things like enemies and in level objects for items not affecting the player directly, also gives more control over changing the player with out messing up the item manger, may be merged into the game directer later down the road.