Fixed bullet class, and rearranged some ui nodes
got some ui ideas down, not sure how to do the weapon switch however. I have a spinning circle idea that could be neat but am not sure how to implement it, maybe just a cicle sprite on a node that has its rot changing that spins 3 other sprites that are children of the circle could work but am still not sure how to fully implement. I am going with 3 weapons being held by the player, seems like a good middle ground of not too many weapons but not limiting either, also might wanna change the weapon switching to swap between 3 different objects containers to stream line and prevent any bugs that can occur with an array that can be what ever number of objects long, probably will be done later tho.